+++ include/System/Animation.hpp
@@ -8,7 +8,7 @@ class Animation
 public:
     Animation();
     Animation(uint32_t sheetW, uint32_t sheetH, uint32_t spriteW, uint32_t spriteH, uint32_t frameCount, float speed);
-    void setTextureID(uint32_t textureID);
+    void setTextureID(uint64_t textureID);
     void setDone() {_done=true;}
     void setCenter(sf::Vector2f center) {_spriteCenter=center;}
     void resetTime(float time);
@@ -16,8 +16,8 @@ public:
     sf::IntRect  getTexCoord(float time);
     sf::Vector2f getSpriteSize()   const {return sf::Vector2f(static_cast<float>(_spriteW), static_cast<float>(_spriteH));}
     sf::Vector2f getSpriteCenter() const {return _spriteCenter;}
-	uint32_t       getTexture()      const {return _textureID;}
-	uint32_t       getCurrentIndex(float time) const {return static_cast<size_t>(_animationSpeed*(time-_startTime));}
+	uint64_t       getTexture()      const {return _textureID;}
+	uint64_t       getCurrentIndex(float time) const {return static_cast<size_t>(_animationSpeed*(time-_startTime));}
 
     void applyOnQuad(sf::VertexArray& quad, float time);
 
@@ -36,7 +36,7 @@ private:
     bool   _done;
 
     sf::Vector2f _spriteCenter;
-	uint32_t       _textureID;
+	uint64_t     _textureID;
 
 };
